+++ System/Entity.h
@@ -382,7 +382,13 @@ namespace RTE {
 		/// Adds this Entity to a new grouping.
 		/// </summary>
 		/// <param name="newGroup">A string which describes the group to add this to. Duplicates will be ignored.</param>
-		void AddToGroup(std::string newGroup) { m_Groups.push_back(newGroup); m_Groups.sort(); m_Groups.unique(); m_LastGroupSearch.clear(); }
+		void AddToGroup(const std::string &newGroup) { m_Groups.push_back(newGroup); m_Groups.sort(); m_Groups.unique(); m_LastGroupSearch.clear(); }
+
+		/// <summary>
+		/// Removes this Entity from the specified grouping.
+		/// </summary>
+		/// <param name="groupToRemoveFrom">A string which describes the group to remove this from.</param>
+		void RemoveFromGroup(const std::string &groupToRemoveFrom) { m_Groups.remove(groupToRemoveFrom); m_LastGroupSearch.clear(); }
 
 		/// <summary>
 		/// Returns random weight used in PresetMan::GetRandomBuyableOfGroupFromTech.